+++ lib/src/document.dart
@@ -20,9 +20,11 @@ class Document extends MapBase<String, dynamic> {
   String get id => _map["_id"];
   set id(String v) => _map["_id"] = v;
 
+  Function onChanged;
+
   /// Create a Document. Optionally include a map of type
   /// Map<String, dynamic> to initially populate the Document with data.
-  Document([Map<String, dynamic> initialValues]) {
+  Document({Map<String, dynamic> initialValues, this.onChanged}) {
     // initial values if provided
     if (initialValues != null) {
       initialValues.keys.forEach((String key) {
@@ -42,12 +44,20 @@ class Document extends MapBase<String, dynamic> {
     save();
   }
 
+  _notifyListener() {
+    if (onChanged != null) {
+      onChanged(this);
+    }
+  }
+
   void clear() {
     _map.clear();
+    _notifyListener();
   }
 
   void remove(Object key) {
     _map.remove(key);
+    _notifyListener();
   }
 
   List<String> get keys {
@@ -58,7 +68,9 @@ class Document extends MapBase<String, dynamic> {
     final file = await _localFile(_map["_id"]);
     // Write the file
     String mapString = json.encode(_map);
-    return file.writeAsString('$mapString');
+    file.writeAsString('$mapString');
+    _notifyListener();
+    return file;
   }
 
   Future<File> _localFile(String id) async {
@@ -88,6 +100,7 @@ class Document extends MapBase<String, dynamic> {
       // even called
       // print("Warning: ${f.path} file was empty.");
     }
+    _notifyListener();
   }
 
   static String randomFileSafeId(int length) {